requirements.txt
langchain
langchain-community
langchain-ollama
chainlit==1.2.0
pyyaml
pytest
structlog